update to dev

+++ modules/Settings/AutomaticAssignment/models/Module.php
@@ -1,46 +1,52 @@
 <?php
 
 /**
- * Automatic assignment module model class
- * @package YetiForce.Settings.Model
- * @license licenses/License.html
+ * Automatic assignment module model class.
+ *
+ * @copyright YetiForce Sp. z o.o
+ * @license YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author Radosław Skrzypczak <r.skrzypczak@yetiforce.com>
  */
 class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_Model
 {
-
 	/**
-	 * Table name
-	 * @var string 
+	 * Table name.
+	 *
+	 * @var string
 	 */
 	public $baseTable = 's_#__automatic_assignment';
 
 	/**
-	 * Table primary key
-	 * @var string 
+	 * Table primary key.
+	 *
+	 * @var string
 	 */
 	public $baseIndex = 'id';
 
 	/**
-	 * List of fields displayed in list view
-	 * @var string 
+	 * List of fields displayed in list view.
+	 *
+	 * @var string
 	 */
 	public $listFields = ['tabid' => 'FL_MODULE', 'field' => 'FL_FIELD', 'value' => 'FL_VALUE', 'active' => 'FL_ACTIVE', 'roleid' => 'FL_MODE'];
 
 	/**
-	 * Module Name
-	 * @var string 
+	 * Module Name.
+	 *
+	 * @var string
 	 */
 	public $name = 'AutomaticAssignment';
 
 	/**
-	 * List of available field types
-	 * @var string[] 
+	 * List of available field types.
+	 *
+	 * @var string[]
 	 */
 	private static $fieldType = ['string'];
 
 	/**
-	 * Function to get the url for Create view of the module
+	 * Function to get the url for Create view of the module.
+	 *
 	 * @return string - url
 	 */
 	public function getCreateRecordUrl()
@@ -49,7 +55,8 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function to get the url for edit view of the module
+	 * Function to get the url for edit view of the module.
+	 *
 	 * @return string - url
 	 */
 	public function getEditViewUrl()
@@ -58,7 +65,8 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function to get the url for default view of the module
+	 * Function to get the url for default view of the module.
+	 *
 	 * @return string URL
 	 */
 	public function getDefaultUrl()
@@ -67,7 +75,8 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function get supported modules
+	 * Function get supported modules.
+	 *
 	 * @return array - List of modules
 	 */
 	public static function getSupportedModules()
@@ -76,7 +85,8 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * List of supported module fields
+	 * List of supported module fields.
+	 *
 	 * @return array
 	 */
 	public static function getFieldsByModule($moduleName)
@@ -92,9 +102,11 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function verifies if it is possible to sort by given field in list view
+	 * Function verifies if it is possible to sort by given field in list view.
+	 *
 	 * @param string $fieldName
-	 * @return boolean
+	 *
+	 * @return bool
 	 */
 	public function isSortByName($fieldName)
 	{
@@ -105,8 +117,9 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function returns list of fields available in list view
-	 * @return Vtiger_Base_Model[]
+	 * Function returns list of fields available in list view.
+	 *
+	 * @return \App\Base[]
 	 */
 	public function getListFields()
 	{
@@ -114,7 +127,7 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 			$fields = $this->listFields;
 			$fieldObjects = [];
 			foreach ($fields as $fieldName => $fieldLabel) {
-				$fieldObject = new Vtiger_Base_Model(['name' => $fieldName, 'label' => $fieldLabel]);
+				$fieldObject = new \App\Base(['name' => $fieldName, 'label' => $fieldLabel]);
 				if (!$this->isSortByName($fieldName)) {
 					$fieldObject->set('sort', true);
 				}
@@ -126,9 +139,11 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function searches for record from the Auto assign records panel
+	 * Function searches for record from the Auto assign records panel.
+	 *
 	 * @param Vtiger_Record_Model $recordModel
-	 * @param string $role
+	 * @param string              $role
+	 *
 	 * @return bool|Settings_AutomaticAssignment_Record_Model
 	 */
 	public function searchRecord(\Vtiger_Record_Model $recordModel, $role = '')
@@ -157,6 +172,7 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 			if ($row['value'] == $recordModel->get($row['field'])) {
 				$autoAssignRecordModel = Settings_AutomaticAssignment_Record_Model::getInstanceById($row['id']);
 				$autoAssignRecordModel->sourceRecordModel = $recordModel;
+
 				return $autoAssignRecordModel;
 			}
 		}
@@ -164,12 +180,13 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Function clears cache
+	 * Function clears cache.
+	 *
 	 * @param array $param
 	 */
 	public function clearCache($param)
 	{
-		if ($param) {
+		if (!empty($param) && isset($param['tabid'], $param['roleid'])) {
 			$tabId = \App\Module::getModuleName($param['tabid']);
 			$cacheKey = $tabId . $param['roleid'];
 			\App\Cache::delete(get_class() . '::searchRecord', $cacheKey);
@@ -177,7 +194,8 @@ class Settings_AutomaticAssignment_Module_Model extends Settings_Vtiger_Module_M
 	}
 
 	/**
-	 * Execute auto assign 
+	 * Execute auto assign.
+	 *
 	 * @param \Vtiger_Record_Model $recordModel
 	 */
 	public static function autoAssignExecute(\Vtiger_Record_Model $recordModel)
