+++ modules/Settings/ModuleManager/models/Library.php
@@ -1,36 +1,37 @@
 <?php
 
 /**
- * Module Manager Library class
- * @package YetiForce.Model
- * @license licenses/License.html
+ * Module Manager Library class.
+ *
+ * @copyright YetiForce Sp. z o.o
+ * @license YetiForce Public License 3.0 (licenses/LicenseEN.txt or yetiforce.com)
  * @author Mariusz Krzaczkowski <m.krzaczkowski@yetiforce.com>
  */
 class Settings_ModuleManager_Library_Model
 {
-
 	/**
-	 * List of all installation libraries
-	 * @var array 
+	 * List of all installation libraries.
+	 *
+	 * @var array
 	 */
 	public static $libraries = [
-		'mPDF' => ['dir' => 'libraries/mPDF/', 'url' => 'https://github.com/YetiForceCompany/lib_mPDF', 'name' => 'lib_mPDF'],
-		'roundcube' => ['dir' => 'modules/OSSMail/roundcube/', 'url' => 'https://github.com/YetiForceCompany/lib_roundcube', 'name' => 'lib_roundcube'],
-		'PHPExcel' => ['dir' => 'libraries/PHPExcel/', 'url' => 'https://github.com/YetiForceCompany/lib_PHPExcel', 'name' => 'lib_PHPExcel'],
-		'AJAXChat' => ['dir' => 'libraries/AJAXChat/', 'url' => 'https://github.com/YetiForceCompany/lib_AJAXChat', 'name' => 'lib_AJAXChat'],
-		'Gantt' => ['dir' => 'libraries/gantt/', 'url' => 'https://github.com/YetiForceCompany/lib_gantt', 'name' => 'lib_gantt'],
+		'mPDF' => ['dir' => 'vendor/mPDF/', 'url' => 'https://github.com/YetiForceCompany/lib_mPDF', 'name' => 'lib_mPDF'],
+		'roundcube' => ['dir' => 'public_html/modules/OSSMail/roundcube/', 'url' => 'https://github.com/YetiForceCompany/lib_roundcube', 'name' => 'lib_roundcube'],
 	];
 
 	/**
-	 * Path to save temporary files
-	 * @var string 
+	 * Path to save temporary files.
+	 *
+	 * @var string
 	 */
 	const TEMP_DIR = 'cache' . DIRECTORY_SEPARATOR . 'upload';
 
 	/**
-	 * Function to check library status
+	 * Function to check library status.
+	 *
 	 * @param string $name
-	 * @return boolean
+	 *
+	 * @return bool
 	 */
 	public static function checkLibrary($name)
 	{
@@ -48,11 +49,13 @@ class Settings_ModuleManager_Library_Model
 			}
 		}
 		App\Cache::save('LIBRARY', $name, $status, App\Cache::LONG);
+
 		return $status;
 	}
 
 	/**
-	 * Get a list of all libraries and their statuses
+	 * Get a list of all libraries and their statuses.
+	 *
 	 * @return array
 	 */
 	public static function &getAll()
@@ -76,8 +79,7 @@ class Settings_ModuleManager_Library_Model
 	}
 
 	/**
-	 * Download all missing libraries
-	 * @throws \Exception\NoPermitted
+	 * Download all missing libraries.
 	 */
 	public static function downloadAll()
 	{
@@ -87,38 +89,36 @@ class Settings_ModuleManager_Library_Model
 	}
 
 	/**
-	 * Function to download library
+	 * Function to download library.
+	 *
 	 * @param string $name
-	 * @return boolean
+	 *
 	 * @throws \Exception\NoPermitted
+	 *
+	 * @return bool
 	 */
 	public static function download($name)
 	{
 		if (!static::$libraries[$name]) {
 			App\Log::warning('Library does not exist: ' . $name);
-			throw new \Exception\NoPermitted('LBL_PERMISSION_DENIED');
+			throw new \App\Exceptions\NoPermitted('LBL_PERMISSION_DENIED', 406);
 		}
-
 		$lib = static::$libraries[$name];
-		if (file_exists($lib['dir'] . 'version.php')) {
-			App\Log::info('Library has already been downloaded: ' . $name);
-			return false;
-		}
-		$path = static::TEMP_DIR . DIRECTORY_SEPARATOR . $name . '.zip';
+		$path = static::TEMP_DIR . DIRECTORY_SEPARATOR . $lib['name'] . '.zip';
 		$mode = AppConfig::developer('MISSING_LIBRARY_DEV_MODE') ? 'developer' : App\Version::get($lib['name']);
 		$compressedName = $lib['name'] . '-' . $mode;
-		if (!file_exists($path)) {
+		if (!file_exists($path) && \App\RequestUtil::isNetConnection()) {
 			stream_context_set_default([
 				'ssl' => [
-					'verify_peer' => false,
-					'verify_peer_name' => false,
+					'verify_peer' => true,
+					'verify_peer_name' => true,
 				],
 			]);
 			$url = $lib['url'] . "/archive/$mode.zip";
 			$headers = get_headers($url, 1);
 			if (isset($headers['Status']) && strpos($headers['Status'], '302') !== false) {
 				App\Log::trace('Started downloading library: ' . $name);
-				if ($file = file_get_contents($url, false, stream_context_create(['ssl' => ['verify_peer' => false, 'verify_peer_name' => false]]))) {
+				if ($file = file_get_contents($url, false, stream_context_create(['ssl' => ['verify_peer' => true, 'verify_peer_name' => true]]))) {
 					file_put_contents($path, $file);
 					App\Log::trace('Completed downloads library: ' . $name);
 				}
@@ -127,9 +127,9 @@ class Settings_ModuleManager_Library_Model
 			}
 		}
 		if (file_exists($path) && filesize($path) > 0) {
-			$unzip = new \vtlib\Unzip($path);
-			$unzip->unzipAllEx('.', [], [$compressedName => $lib['dir']]);
-			$unzip->close();
+			\vtlib\Functions::recurseDelete($lib['dir']);
+			$zip = \App\Zip::openFile($path, ['checkFiles' => false]);
+			$zip->unzip([$compressedName => $lib['dir']]);
 			unlink($path);
 		} else {
 			App\Log::warning('No import file: ' . $name);
@@ -137,14 +137,12 @@ class Settings_ModuleManager_Library_Model
 	}
 
 	/**
-	 * Function to update library
+	 * Function to update library.
+	 *
 	 * @param string $name
-	 * @throws \Exception\NoPermitted
 	 */
 	public static function update($name)
 	{
-		$lib = static::$libraries[$name];
-		\vtlib\Functions::recurseDelete($lib['dir']);
 		static::download($name);
 	}
 }
